pyproject.toml
[tool.poetry]
name = "tlp-project"
version = "v2025.12.0"
description = "Product Market Share Analysis - ML tool for predicting cargo product market shares per O&D route"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
aiofiles = "24.1.0"
colorama = "0.4.6"
dict-toolbox = "5.0.0"
Jinja2 = "3.1.4"
joblib = "1.4.2"
lazy-object-proxy = "1.10.0"
MarkupSafe = "3.0.2"
msgpack = "1.1.0"
nest-asyncio = "1.6.0"
numpy = "2.0.2"
pandas = "2.2.3"
pop = "27.1.0"
pop-config = "12.0.4"
pop-loop = "1.1.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "0.21.0"
pytz = "2024.2"
PyYAML = "6.0.2"
rend = "7.0.2"
scikit-learn = "1.5.2"
scipy = "1.13.1"
six = "1.16.0"
sniffio = "1.3.1"
threadpoolctl = "3.5.0"
toml = "0.10.2"
tzdata = "2024.2"
flask = "*"
gcsfs = "*"
lightgbm = ">=3.3"

[tool.poetry.group.dev.dependencies]
pytest = "8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

